requirements.txt
# RAG System Requirements
# Python 3.8+

# Core RAG Framework
langchain==0.1.0
langchain-community==0.0.13
langchain-google-genai==0.0.6

# Google Gemini API
google-generativeai==0.3.2

# PDF Processing
pypdf==3.17.4

# Embeddings
sentence-transformers==2.2.2

# Vector Database
faiss-cpu==1.7.4

# Web UI
streamlit==1.29.0

# Additional Dependencies
numpy>=1.24.0
pandas>=2.0.0
